# src/server/api/go/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Acontext.",
        "title": "Acontext API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/tool/name": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all tool names within a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tool"
                ],
                "summary": "Get tool names",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/httpclient.ToolReferenceData"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Rename one or more tool names within a project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tool"
                ],
                "summary": "Rename tool names",
                "parameters": [
                    {
                        "description": "Tool rename request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.RenameToolNameReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/serializer.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/httpclient.FlagResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.RenameToolNameReq": {
            "type": "object",
            "required": [
                "rename"
            ],
            "properties": {
                "rename": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/handler.ToolRenameItem"
                    }
                }
            }
        },
...
        "handler.ToolRenameItem": {
            "type": "object",
            "required": [
                "new_name",
                "old_name"
            ],
            "properties": {
                "new_name": {
                    "type": "string"
                },
                "old_name": {
                    "type": "string"
                }
            }
        },
...
        "httpclient.FlagResponse": {
            "type": "object",
            "properties": {
                "errmsg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "httpclient.ToolReferenceData": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "sop_count": {
                    "type": "integer"
                }
            }
        },
...
        "serializer.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Project Bearer token (e.g., \"Bearer sk-ac-xxxx\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}